requirements.txt
alembic==1.14.0
Flask==3.1.0
Flask_Cors==5.0.0
Flask_JWT_Extended==4.7.1
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
SQLAlchemy==2.0.36
Werkzeug==3.1.3
bcrypt
itsdangerous
pymysql
cryptography